begin refactor of common code
