+++ src/common.js
@@ -1,0 +1,7 @@
+/**
+ * Manually add common modules for directives
+ */
+export { isAttributePart, isNodePart } from './parts.js';
+export { nothingString, unsafePrefixString } from './string.js';
+export { directive } from './directive.js';
+export { isPrimitive } from './is.js';
