+++ src/directives/async-append.js
@@ -1,4 +1,4 @@
-import { directive, isNodePart } from '../index.js';
+import { directive, isNodePart } from '../common.js';
 
 /**
  * Render items of an AsyncIterable
